+++ index.ts
@@ -23,30 +23,32 @@ export function style({ minify = true, charset = "utf8" }: StylePluginOptions = 
   return {
     name: "style",
     setup({ onResolve, onLoad, esbuild }) {
-      const cwd = process.cwd();
       const opt: BuildOptions = { logLevel: "silent", bundle: true, write: false, charset, minify };
       const require_esbuild = () => esbuild || (esbuild_shim ||= require("esbuild"));
 
-      onResolve({ filter: /\.css$/, namespace: "file" }, args => {
-        const absPath = path.join(args.resolveDir, args.path);
-        const relPath = path.relative(cwd, absPath);
-        const resolved = fs.existsSync(absPath) ? relPath : args.path;
-        return { path: resolved, namespace: "style-stub" };
+      onLoad({ filter: /\.css$/ }, async args => {
+        const { errors, warnings, outputFiles } = await require_esbuild().build({
+          entryPoints: [args.path],
+          ...opt,
+        });
+        const css = outputFiles![0].text.trimEnd();
+        return {
+          errors,
+          warnings,
+          contents: `
+            import { inject_style } from "__style_helper__"
+            inject_style(${JSON.stringify(css)})
+          `,
+        };
       });
 
-      onResolve({ filter: /\.css$/, namespace: "style-stub" }, args => {
-        return { path: args.path, namespace: "style-content" };
+      onResolve({ filter: /^__style_helper__$/ }, () => {
+        return { path: "index.js", namespace: "style-helper" };
       });
 
-      onResolve({ filter: /^__style_helper__$/, namespace: "style-stub" }, args => ({
-        path: args.path,
-        namespace: "style-helper",
-        sideEffects: false,
-      }));
-
-      onLoad({ filter: /.*/, namespace: "style-helper" }, async () => ({
+      onLoad({ filter: /.*/, namespace: "style-helper" }, () => ({
         contents: `
-          export function injectStyle(text) {
+          export function inject_style(text) {
             if (typeof document !== 'undefined') {
               var style = document.createElement('style')
               var node = document.createTextNode(text)
@@ -56,20 +58,6 @@ export function style({ minify = true, charset = "utf8" }: StylePluginOptions = 
           }
         `,
       }));
-
-      onLoad({ filter: /.*/, namespace: "style-stub" }, async args => ({
-        contents: `
-          import { injectStyle } from "__style_helper__"
-          import css from ${JSON.stringify(args.path)}
-          injectStyle(css)
-        `,
-      }));
-
-      onLoad({ filter: /.*/, namespace: "style-content" }, async args => {
-        const options = { entryPoints: [args.path], ...opt };
-        const { errors, warnings, outputFiles } = await require_esbuild().build(options);
-        return { errors, warnings, contents: outputFiles![0].text, loader: "text" };
-      });
     },
   };
 }